% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forward_Select.R
\name{forward_select}
\alias{forward_select}
\title{Forward Selection in Linear Regression}
\usage{
forward_select(data, response)
}
\arguments{
\item{data}{The data set containing the variables.}

\item{response}{The response variable in the model.}
}
\value{
a fitted linear model
}
\description{
This function implements a forward selection algorithm to build
a linear regression model by sequentially adding predictors that minimize the AIC.
}
\examples{
data(mtcars)
model <- forward_select(mtcars, mpg)
summary(model)


}
